refactor: move cli to own file
